Updated Data Types based on understanding of the Instructions

+++ src/main/kotlin/activity_01_d/Main.kt
@@ -14,7 +14,7 @@ fun main() {
     var value4: Double? = null
     var value5: Double? = null
     var divisor: Int? = null
-    var sum: Int = 0
+    var sum: Double = 0.0
     var quotient: Double? = null
 
     // Value Entry
